:sparkles: DOP-4240 implements filtering sensitive values method to prevent sensitive values being logged
